functionfs: Work around kernel bug on f_fs teardown with eventfd.
The kernel has an underflow bug which gets triggered during gadget
teardown when an eventfd was provided to f_fs. Detect (with a simplistic
heuristic) affected versions. On such versions, an AIO block is used and
given the same eventfd, so code using this module sees no difference.
Internally, the difference is that EP0 event processing happens during
the AIO block completion callback, which is called during processEvents,
instead of by processEvents itself. This has the overhead of re-submitting
the AIO block each time.
Default heuristic only sets a cut-off at 5.16, first major kernel release
containing the fix, as accessing minor versions seems hard to do (the
full version string in uname output is burried in the middle of
distro-dependent strings).