WIP functionfs: Add support for systemd.socket-style gadget initialisation
WIP: untested, subject to changes

See (5)systemd.socket, ListenUSBFunction.
In such setup, systemd opens endpoint zero, writes the descriptors and
opens all endpoints. As a result:
- endpoints are received as already-open file descriptors
- neither function nor string descriptors may be written to endpoint 0
  upon function activation
Design decision: requiring the developer to provide a python representation
of the function, which must always consistent with the descriptor provided
to the kernel (which must be stored in a file prior to the function process
starting) would only cause confusion. Instead, expect to receive the
content of the file containing the function descriptor as received by
systemd, and parse it to get the pieces of information necessary for event
monitoring.
All this is done without adding any dependency on systemd code, only making
it possible to use in such context.